It compiles

+++ src/Yale/Parser/ExpressionTokenizer.cs
@@ -17,125 +17,100 @@ internal sealed class ExpressionTokenizer : Tokenizer
 
     private void CreatePatterns()
     {
-        TokenPattern pattern = new((int)TokenId.ADD, "ADD", PatternType.STRING, "+");
+        TokenPattern pattern = new(TokenId.ADD, "ADD", PatternType.STRING, "+");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.SUB, "SUB", PatternType.STRING, "-");
+        pattern = new TokenPattern(TokenId.SUB, "SUB", PatternType.STRING, "-");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.MUL, "MUL", PatternType.STRING, "*");
+        pattern = new TokenPattern(TokenId.MUL, "MUL", PatternType.STRING, "*");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.DIV, "DIV", PatternType.STRING, "/");
+        pattern = new TokenPattern(TokenId.DIV, "DIV", PatternType.STRING, "/");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.POWER, "POWER", PatternType.STRING, "^");
+        pattern = new TokenPattern(TokenId.POWER, "POWER", PatternType.STRING, "^");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.MOD, "MOD", PatternType.STRING, "%");
+        pattern = new TokenPattern(TokenId.MOD, "MOD", PatternType.STRING, "%");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.LEFT_PAREN, "LEFT_PAREN", PatternType.STRING, "(");
+        pattern = new TokenPattern(TokenId.LEFT_PAREN, "LEFT_PAREN", PatternType.STRING, "(");
         AddPattern(pattern);
 
-        pattern = new TokenPattern(
-            (int)TokenId.RIGHT_PAREN,
-            "RIGHT_PAREN",
-            PatternType.STRING,
-            ")"
-        );
+        pattern = new TokenPattern(TokenId.RIGHT_PAREN, "RIGHT_PAREN", PatternType.STRING, ")");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.LEFT_BRACE, "LEFT_BRACE", PatternType.STRING, "[");
+        pattern = new TokenPattern(TokenId.LEFT_BRACE, "LEFT_BRACE", PatternType.STRING, "[");
         AddPattern(pattern);
 
-        pattern = new TokenPattern(
-            (int)TokenId.RIGHT_BRACE,
-            "RIGHT_BRACE",
-            PatternType.STRING,
-            "]"
-        );
+        pattern = new TokenPattern(TokenId.RIGHT_BRACE, "RIGHT_BRACE", PatternType.STRING, "]");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.EQ, "EQ", PatternType.STRING, "=");
+        pattern = new TokenPattern(TokenId.EQ, "EQ", PatternType.STRING, "=");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.EQ, "EQ", PatternType.STRING, "EQ");
+        pattern = new TokenPattern(TokenId.EQ, "EQ", PatternType.STRING, "EQ");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.LT, "LT", PatternType.STRING, "<");
+        pattern = new TokenPattern(TokenId.LT, "LT", PatternType.STRING, "<");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.GT, "GT", PatternType.STRING, ">");
+        pattern = new TokenPattern(TokenId.GT, "GT", PatternType.STRING, ">");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.LTE, "LTE", PatternType.STRING, "<=");
+        pattern = new TokenPattern(TokenId.LTE, "LTE", PatternType.STRING, "<=");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.GTE, "GTE", PatternType.STRING, ">=");
+        pattern = new TokenPattern(TokenId.GTE, "GTE", PatternType.STRING, ">=");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.NE, "NE", PatternType.STRING, "<>");
+        pattern = new TokenPattern(TokenId.NE, "NE", PatternType.STRING, "<>");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.AND, "AND", PatternType.STRING, "AND");
+        pattern = new TokenPattern(TokenId.AND, "AND", PatternType.STRING, "AND");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.OR, "OR", PatternType.STRING, "OR");
+        pattern = new TokenPattern(TokenId.OR, "OR", PatternType.STRING, "OR");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.XOR, "XOR", PatternType.STRING, "XOR");
+        pattern = new TokenPattern(TokenId.XOR, "XOR", PatternType.STRING, "XOR");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.NOT, "NOT", PatternType.STRING, "NOT");
+        pattern = new TokenPattern(TokenId.NOT, "NOT", PatternType.STRING, "NOT");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.IN, "IN", PatternType.STRING, "in");
+        pattern = new TokenPattern(TokenId.IN, "IN", PatternType.STRING, "in");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.DOT, "DOT", PatternType.STRING, ".");
+        pattern = new TokenPattern(TokenId.DOT, "DOT", PatternType.STRING, ".");
         AddPattern(pattern);
 
         pattern = new TokenPattern(
-            (int)TokenId.ARGUMENT_SEPARATOR,
+            TokenId.ARGUMENT_SEPARATOR,
             "ARGUMENT_SEPARATOR",
             PatternType.STRING,
             ";"
         );
         AddPattern(pattern);
 
-        pattern = new TokenPattern(
-            (int)TokenId.ARRAY_BRACES,
-            "ARRAY_BRACES",
-            PatternType.STRING,
-            "[]"
-        );
+        pattern = new TokenPattern(TokenId.ARRAY_BRACES, "ARRAY_BRACES", PatternType.STRING, "[]");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.LEFT_SHIFT, "LEFT_SHIFT", PatternType.STRING, "<<");
+        pattern = new TokenPattern(TokenId.LEFT_SHIFT, "LEFT_SHIFT", PatternType.STRING, "<<");
         AddPattern(pattern);
 
-        pattern = new TokenPattern(
-            (int)TokenId.RIGHT_SHIFT,
-            "RIGHT_SHIFT",
-            PatternType.STRING,
-            ">>"
-        );
+        pattern = new TokenPattern(TokenId.RIGHT_SHIFT, "RIGHT_SHIFT", PatternType.STRING, ">>");
         AddPattern(pattern);
 
-        pattern = new TokenPattern(
-            (int)TokenId.WHITESPACE,
-            "WHITESPACE",
-            PatternType.REGEXP,
-            "\\s+"
-        )
+        pattern = new TokenPattern(TokenId.WHITESPACE, "WHITESPACE", PatternType.REGEXP, "\\s+")
         {
             Ignore = true
         };
         AddPattern(pattern);
 
         pattern = new TokenPattern(
-            (int)TokenId.INTEGER,
+            TokenId.INTEGER,
             "INTEGER",
             PatternType.REGEXP,
             "\\d+(u|l|ul|lu|f|m)?"
@@ -143,7 +118,7 @@ internal sealed class ExpressionTokenizer : Tokenizer
         AddPattern(pattern);
 
         pattern = new TokenPattern(
-            (int)TokenId.REAL,
+            TokenId.REAL,
             "REAL",
             PatternType.REGEXP,
             "\\d*\\.\\d+([e][+-]\\d{1,3})?(d|f|m)?"
@@ -151,7 +126,7 @@ internal sealed class ExpressionTokenizer : Tokenizer
         AddPattern(pattern);
 
         pattern = new TokenPattern(
-            (int)TokenId.STRING_LITERAL,
+            TokenId.STRING_LITERAL,
             "STRING_LITERAL",
             PatternType.REGEXP,
             "\"([^\"\\r\\n\\\\]|\\\\u[0-9a-f]{4}|\\\\[\\\\\"'trn])*\""
@@ -159,21 +134,21 @@ internal sealed class ExpressionTokenizer : Tokenizer
         AddPattern(pattern);
 
         pattern = new TokenPattern(
-            (int)TokenId.CHAR_LITERAL,
+            TokenId.CHAR_LITERAL,
             "CHAR_LITERAL",
             PatternType.REGEXP,
             "'([^'\\r\\n\\\\]|\\\\u[0-9a-f]{4}|\\\\[\\\\\"'trn])'"
         );
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.TRUE, "TRUE", PatternType.STRING, "True");
+        pattern = new TokenPattern(TokenId.TRUE, "TRUE", PatternType.STRING, "True");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.FALSE, "FALSE", PatternType.STRING, "False");
+        pattern = new TokenPattern(TokenId.FALSE, "FALSE", PatternType.STRING, "False");
         AddPattern(pattern);
 
         pattern = new TokenPattern(
-            (int)TokenId.IDENTIFIER,
+            TokenId.IDENTIFIER,
             "IDENTIFIER",
             PatternType.REGEXP,
             "[a-z_]\\w*"
@@ -181,7 +156,7 @@ internal sealed class ExpressionTokenizer : Tokenizer
         AddPattern(pattern);
 
         pattern = new TokenPattern(
-            (int)TokenId.HEX_LITERAL,
+            TokenId.HEX_LITERAL,
             "HEX_LITERAL",
             PatternType.REGEXP,
             "0x[0-9a-f]+(u|l|ul|lu)?"
@@ -189,7 +164,7 @@ internal sealed class ExpressionTokenizer : Tokenizer
         AddPattern(pattern);
 
         pattern = new TokenPattern(
-            (int)TokenId.NULL_LITERAL,
+            TokenId.NULL_LITERAL,
             "NULL_LITERAL",
             PatternType.STRING,
             "null"
@@ -197,25 +172,20 @@ internal sealed class ExpressionTokenizer : Tokenizer
         AddPattern(pattern);
 
         pattern = new TokenPattern(
-            (int)TokenId.TIMESPAN,
+            TokenId.TIMESPAN,
             "TIMESPAN",
             PatternType.REGEXP,
             "##(\\d+\\.)?\\d{2}:\\d{2}(:\\d{2}(\\.\\d{1,7})?)?#"
         );
         AddPattern(pattern);
 
-        pattern = new TokenPattern(
-            (int)TokenId.DATETIME,
-            "DATETIME",
-            PatternType.REGEXP,
-            "#[^#]+#"
-        );
+        pattern = new TokenPattern(TokenId.DATETIME, "DATETIME", PatternType.REGEXP, "#[^#]+#");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.IF, "IF", PatternType.STRING, "if");
+        pattern = new TokenPattern(TokenId.IF, "IF", PatternType.STRING, "if");
         AddPattern(pattern);
 
-        pattern = new TokenPattern((int)TokenId.CAST, "CAST", PatternType.STRING, "cast");
+        pattern = new TokenPattern(TokenId.CAST, "CAST", PatternType.STRING, "cast");
         AddPattern(pattern);
     }
 }
